Obfuscate with some small trickery
First, instead of overwriting argv[0] we just increment argv if
the user passed in an argument by abusing the fact that boolean
expressions are either 0 or 1. This has the same effect as previously
without any overwriting and is more confusing.

Second, we use the XOR trick to zero out argc in the base case.

Third, we get rid of all the offseting by one in the final recursion
statement by using the prefix decrement in the test itself.